# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Maicare server API documentation.",
        "title": "Maicare API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "your-email@domain.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/clients": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "List clients",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Location ID",
                        "name": "location_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-pagination_Response-api_ListClientsApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Create a new client",
                "parameters": [
                    {
                        "description": "Client details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateClientDetailsRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.Response-api_CreateClientDetailsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/clients/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Get a client",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Client ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-api_GetClientApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
...
        "/clients/{id}/profile_picture": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Set a client profile picture",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Client ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Client profile picture",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.SetClientProfilePictureRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-api_SetClientProfilePictureResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.Address": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "belongs_to": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "zip_code": {
                    "type": "string"
                }
            }
        },
...
        "api.CreateClientDetailsRequest": {
            "type": "object",
            "required": [
                "birthplace",
                "date_of_birth",
                "departement",
                "email",
                "filenumber",
                "first_name",
                "gender",
                "last_name",
                "location_id",
                "organisation",
                "phone_number",
                "sender_id",
                "source"
            ],
            "properties": {
                "addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Address"
                    }
                },
                "birthplace": {
                    "type": "string"
                },
                "bsn": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "departement": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "filenumber": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "identity_attachment_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "infix": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "legal_measure": {
                    "type": "string"
                },
                "location_id": {
                    "type": "integer"
                },
                "organisation": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "sender_id": {
                    "type": "integer"
                },
                "source": {
                    "type": "string"
                }
            }
        },
        "api.CreateClientDetailsResponse": {
            "type": "object",
            "properties": {
                "addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Address"
                    }
                },
                "birthplace": {
                    "type": "string"
                },
                "bsn": {
                    "type": "string"
                },
                "created": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "departement": {
                    "type": "string"
                },
                "departure_reason": {
                    "type": "string"
                },
                "departure_report": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "filenumber": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "has_untaken_medications": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "identity": {
                    "type": "boolean"
                },
                "identity_attachment_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "infix": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "legal_measure": {
                    "type": "string"
                },
                "location_id": {
                    "type": "integer"
                },
                "organisation": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "profile_picture": {
                    "type": "string"
                },
                "sender_id": {
                    "type": "integer"
                },
                "source": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "api.Response-any": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.Response-api_CreateClientDetailsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/api.CreateClientDetailsResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.Response-api_GetClientApiResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/api.GetClientApiResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.Response-api_SetClientProfilePictureResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/api.SetClientProfilePictureResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.Response-pagination_Response-api_ListClientsApiResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/pagination.Response-api_ListClientsApiResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.SetClientProfilePictureRequest": {
            "type": "object",
            "required": [
                "attachement_id"
            ],
            "properties": {
                "attachement_id": {
                    "type": "string"
                }
            }
        },
        "api.SetClientProfilePictureResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "profile_picture": {
                    "type": "string"
                }
            }
        },
...
        "pagination.Response-api_ListClientsApiResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "next": {
                    "type": "string"
                },
                "page_size": {
                    "type": "integer"
                },
                "previous": {
                    "type": "string"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ListClientsApiResponse"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Add 'Bearer ' prefix before your JWT token for authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}